requirements.txt
# requirements.txt
Flask==2.2.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
gunicorn==20.1.0
python-dotenv==1.0.0
alembic==1.12.0
pymysql==1.1.0
bcrypt==4.0.1